Tweaked styles and re-added trigger block.
Many small tweaks to sizing and scaling of blocks to fit new style of block.

Trigger block now spawned with coord relative to canvas to allow for screen positioning.

Updated all panel textures and added a couple missing to be sliced as close to the edge as they can to allow full diagonals.

Block supplier now keeps track of z coordinate of original object being instantiated.